+++ src/main/java/uk/ac/standrews/cs/population_linkage/linkageRecipes/BirthBrideIdentityLinkageRecipe.java
@@ -162,7 +162,7 @@ public class BirthBrideIdentityLinkageRecipe extends LinkageRecipe {
     }
 
     @Override
-    public int getNumberOfGroundTruthTrueLinks() {
+    public long getNumberOfGroundTruthTrueLinks() {
         return getNumberOfGroundTruthLinksAsymmetric();
     }
 }
