+++ src/main/java/uk/ac/standrews/cs/population_linkage/linkageRecipes/BirthBrideSiblingLinkageRecipe.java
@@ -138,7 +138,7 @@ public class BirthBrideSiblingLinkageRecipe extends LinkageRecipe {
     }
 
     @Override
-    public int getNumberOfGroundTruthTrueLinks() {
+    public long getNumberOfGroundTruthTrueLinks() {
         return getNumberOfGroundTruthLinksAsymmetric();
     }
