+++ src/main/java/uk/ac/standrews/cs/population_linkage/linkageRecipes/BirthSiblingLinkageRecipe.java
@@ -149,7 +149,7 @@ public class BirthSiblingLinkageRecipe extends LinkageRecipe {
     }
 
     @Override
-    public int getNumberOfGroundTruthTrueLinks() {
+    public long getNumberOfGroundTruthTrueLinks() {
         return getNumberOfGroundTruthLinksSymmetric();
     }
