+++ src/main/java/uk/ac/standrews/cs/population_linkage/linkageRecipes/DeathBrideSiblingLinkageRecipe.java
@@ -141,7 +141,7 @@ public class DeathBrideSiblingLinkageRecipe extends LinkageRecipe {
     }
 
     @Override
-    public int getNumberOfGroundTruthTrueLinks() {
+    public long getNumberOfGroundTruthTrueLinks() {
         return getNumberOfGroundTruthLinksAsymmetric();
     }
